Separate Query from Command resource
