+++ src/main/java/ndr/brt/GameCommandResource.java
@@ -1,0 +1,16 @@
+package ndr.brt;
+
+public class GameCommandResource {
+
+    private CommandHandler commandHandler;
+
+    public GameCommandResource() {
+        this.commandHandler = new CommandHandler();
+    }
+
+    public void roll(int count) {
+        RollCommand command = new RollCommand(count);
+
+        commandHandler.handle(command);
+    }
+}
