+++ src/main/java/ndr/brt/GameQueryResource.java
@@ -1,0 +1,11 @@
+package ndr.brt;
+
+public class GameQueryResource {
+
+    private ScoreRepository scoreRepository = ScoreRepository.getInstance();
+
+    public int score() {
+        return scoreRepository.getScore();
+    }
+
+}
